+++ examples/resources/fmc_access_control_policy_inheritance/import.sh
@@ -1,0 +1,2 @@
+# <domain> is optional. If not provided, `Global` is used implicitly and resource's `domain` attribute is not set.
+terraform import fmc_access_control_policy_inheritance.example "<domain>,<access_control_policy_id>,<id>"