+++ internal/provider/data_source_fmc_access_control_policy.go
@@ -663,6 +663,7 @@ func (d *AccessControlPolicyDataSource) Configure(_ context.Context, req datasou
 
 func (d *AccessControlPolicyDataSource) Read(ctx context.Context, req datasource.ReadRequest, resp *datasource.ReadResponse) {
 	var config AccessControlPolicy
+	var hasInheritance bool = false
 
 	// Read config
 	diags := req.Config.Get(ctx, &config)
@@ -721,10 +722,22 @@ func (d *AccessControlPolicyDataSource) Read(ctx context.Context, req datasource
 
 	// Set string that will have categories and rules injected
 	replace := res.String()
+	policyName := res.Get("name").String()
 
 	// Save state of categories and rules management
 	manageCategories := false
 	manageRules := false
+
+	if (!config.ManageCategories.IsUnknown() && config.ManageCategories.ValueBool()) || (!config.ManageRules.IsUnknown() && config.ManageRules.ValueBool()) {
+		resInh, err := d.client.Get(config.getPath()+"/"+url.QueryEscape(config.Id.ValueString())+"/inheritancesettings?expanded=true", reqMods...)
+		if err != nil {
+			resp.Diagnostics.AddError("Client Error", fmt.Sprintf("Failed to retrieve inheritance settings (GET), got error: %s, %s", err, resInh.String()))
+			return
+		}
+		if resInh.Get("items.0.basePolicy.name").Exists() {
+			hasInheritance = true
+		}
+	}
 
 	// If manage_categories is set to true, retrieve categories
 	if !config.ManageCategories.IsUnknown() && config.ManageCategories.ValueBool() {
@@ -740,6 +753,10 @@ func (d *AccessControlPolicyDataSource) Read(ctx context.Context, req datasource
 		replaceCats := resCats.Get("items").String()
 		if replaceCats == "" {
 			replaceCats = "[]"
+		}
+		// If the rule inheritance is enabled, API will include categories from parent policies in the response
+		if hasInheritance {
+			replaceCats = config.filterByPolicyName(replaceCats, policyName)
 		}
 		replace, _ = sjson.SetRaw(replace, "dummy_categories", replaceCats)
 		manageCategories = true
@@ -759,6 +776,10 @@ func (d *AccessControlPolicyDataSource) Read(ctx context.Context, req datasource
 		replaceRules := resRules.Get("items").String()
 		if replaceRules == "" {
 			replaceRules = "[]"
+		}
+		// If the rule inheritance is enabled, API will include rules from parent policies in the response
+		if hasInheritance {
+			replaceRules = config.filterByPolicyName(replaceRules, policyName)
 		}
 		replace, _ = sjson.SetRaw(replace, "dummy_rules", replaceRules)
 		manageRules = true